# backend/openapi.yaml
openapi: 3.0.1
info:
  title: Crooodle
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /admin/world-regions/{id}:
    get:
      tags:
      - world-region-admin-controller
      operationId: getWorldRegionAdmin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorldRegionAdminResponseDto"
    put:
      tags:
      - world-region-admin-controller
      operationId: editWorldRegion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorldRegionEditRequestDto"
        required: true
      responses:
        "200":
          description: OK
    delete:
      tags:
      - world-region-admin-controller
      operationId: deleteWorldRegion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /admin/countries/{id}:
    get:
      tags:
      - country-admin-controller
      operationId: getCountryAdmin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CountryAdminResponseDto"
    put:
      tags:
      - country-admin-controller
      operationId: editCountry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountryEditRequestDto"
        required: true
      responses:
        "200":
          description: OK
    delete:
      tags:
      - country-admin-controller
      operationId: deleteCountry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccessTokenResponseDto"
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccessTokenResponseDto"
  /admin/world-regions:
    get:
      tags:
      - world-region-admin-controller
      operationId: listWorldRegionsAdmin
      parameters:
      - name: arg0
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/WorldRegionCriteriaDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoWorldRegionAdminResponseDto"
    post:
      tags:
      - world-region-admin-controller
      operationId: createWorldRegion
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorldRegionCreateRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
  /admin/countries:
    get:
      tags:
      - country-admin-controller
      operationId: listCountriesAdmin
      parameters:
      - name: arg0
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CountryCriteriaDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoCountryAdminResponseDto"
    post:
      tags:
      - country-admin-controller
      operationId: createCountry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CountryCreateRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
  /world-regions:
    get:
      tags:
      - world-region-controller
      operationId: listWorldRegions
      parameters:
      - name: arg0
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/WorldRegionCriteriaDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoWorldRegionResponseDto"
  /world-regions/{id}:
    get:
      tags:
      - world-region-controller
      operationId: getWorldRegion
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorldRegionResponseDto"
  /countries:
    get:
      tags:
      - country-controller
      operationId: listCountries
      parameters:
      - name: arg0
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/CountryCriteriaDto"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoCountryResponseDto"
  /countries/{id}:
    get:
      tags:
      - country-controller
      operationId: getCountry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CountryResponseDto"
components:
  schemas:
    WorldRegionEditRequestDto:
      type: object
      properties:
        name:
          maxLength: 500
          minLength: 1
          type: string
        countries:
          type: array
          items:
            type: integer
            format: int32
    CountryEditRequestDto:
      type: object
      properties:
        name:
          maxLength: 300
          minLength: 1
          type: string
        worldRegionId:
          type: integer
          format: int32
    UserRegisterRequestDto:
      required:
      - email
      - name
      - password
      type: object
      properties:
        name:
          maxLength: 2147483647
          minLength: 3
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
    AccessTokenResponseDto:
      type: object
      properties:
        accessToken:
          type: string
    UserLoginRequestDto:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
    WorldRegionCreateRequestDto:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 500
          minLength: 1
          type: string
        countries:
          type: array
          items:
            type: integer
            format: int32
    CountryCreateRequestDto:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 300
          minLength: 1
          type: string
        worldRegionId:
          type: integer
          format: int32
    PaginationDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 0
        limit:
          type: integer
          format: int32
          example: 10
    WorldRegionCriteriaDto:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/PaginationDto"
    CountryResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        worldRegion:
          $ref: "#/components/schemas/WorldRegionResponseDto"
    PageResponseDtoWorldRegionResponseDto:
      type: object
      properties:
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorldRegionResponseDto"
    WorldRegionResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        popularCountries:
          type: array
          items:
            $ref: "#/components/schemas/CountryResponseDto"
    CountryCriteriaDto:
      type: object
      properties:
        query:
          type: string
        pagination:
          $ref: "#/components/schemas/PaginationDto"
    PageResponseDtoCountryResponseDto:
      type: object
      properties:
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/CountryResponseDto"
    PageResponseDtoWorldRegionAdminResponseDto:
      type: object
      properties:
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorldRegionAdminResponseDto"
    WorldRegionAdminResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        countries:
          type: array
          items:
            $ref: "#/components/schemas/CountryResponseDto"
    CountryAdminResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        worldRegion:
          $ref: "#/components/schemas/WorldRegionResponseDto"
    PageResponseDtoCountryAdminResponseDto:
      type: object
      properties:
        total:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/CountryAdminResponseDto"
  securitySchemes:
    Bearer:
      type: http
      scheme: Bearer
      bearerFormat: JWT